refactor: extract TelemetryParams and PrometheusParams
