+++ client/cli/src/params/mod.rs
@@ -25,6 +25,8 @@ mod offchain_worker_params;
 mod pruning_params;
 mod shared_params;
 mod transaction_pool_params;
+mod prometheus_params;
+mod telemetry_params;
 
 use crate::arg_enums::{CryptoScheme, OutputType};
 use clap::Args;
@@ -38,7 +40,7 @@ use std::{fmt::Debug, str::FromStr};
 pub use crate::params::{
 	database_params::*, import_params::*, keystore_params::*, message_params::*, network_params::*,
 	node_key_params::*, offchain_worker_params::*, pruning_params::*, shared_params::*,
-	transaction_pool_params::*,
+	transaction_pool_params::*, prometheus_params::*, telemetry_params::*,
 };
 
 /// Parse Ss58AddressFormat
